Moved classes and related interfaces to another package.
