+++ app/src/main/java/com/astronauts/astronaut/dao/object/Search.java
@@ -1,4 +1,4 @@
-package com.astronauts.astronaut.dao;
+package com.astronauts.astronaut.dao.object;
 
 /**
  * Created by end on 01.05.17.
